+++ complexplorer/plotting/__init__.py
@@ -1,5 +1,1 @@
 """Plotting modules for complexplorer."""
-
-from .base import *
-
-__all__ = ["base"]